refactor: removed unused updateDocument method and refactored service
