+++ electron/handlers/core-handler.ts
@@ -3,7 +3,6 @@ import { verifyLangfuseConfig } from "../services/observability/verify-langfuse-
 import { getSuggestions } from "../api/core/get-suggestions";
 import { getAppConfig } from "../api/core/get-app-config";
 import { chatWithAI } from "../api/core/chat";
-import { updateDocument } from "../api/core/document-update";
 import { inlineEditWithAI } from "../api/core/inline-edit";
 import { ipcMain, IpcMainInvokeEvent } from "electron";
 
@@ -54,16 +53,6 @@ export function setupCoreHandlers() {
       return result;
     } catch (error: any) {
       console.error('Error handling core:verifyLangfuseConfig:', error.message);
-      throw error;
-    }
-  });
-
-  ipcMain.handle('core:updateDocument', async (_event: IpcMainInvokeEvent, data: any) => {
-    try {
-      const result = await updateDocument(_event, data);
-      return result;
-    } catch (error: any) {
-      console.error('Error handling core:updateDocument:', error.message);
       throw error;
     }
   });
